requirements.txt
PyGithub>=2.5.0
requests>=2.32.0
tenacity>=9.0.0
tabulate>=0.9.0
charset-normalizer>=3.3.2
PyYAML>=6.0.2
